+++ tools/libexec/perl/TTP/Meteor.pm
@@ -35,6 +35,13 @@ use TTP::Constants qw( :all );
 use TTP::Message qw( :all );
 
 my $Const = {
+	package => {
+		md => [
+			'ChangeLog.md',
+			'README.md',
+			'TODO.md'
+		]
+	}
 };
 
 # ------------------------------------------------------------------------------------------------
@@ -88,6 +95,163 @@ sub getApplication {
 	}
 
 	return $res;
+}
+
+# ------------------------------------------------------------------------------------------------
+# does the given directory host a Meteor package ?
+# (I):
+# - a directory path
+# (O):
+# - a package object with following keys:
+#   > dir: the path to the package directory
+#   > js: the path to the package.js file
+#   > name: the name of the package
+#   > ChangeLog.md: the full path to ChangeLog.md
+#   > README.md: the full path to README.md
+#   > TODO.md: the full path to TODO.md
+# - returns undef if not a Meteor package
+
+sub getPackage {
+	my ( $dir ) = @_;
+	my $res = undef;
+
+	if( -d $dir ){
+		# test package.js
+		my $jspck = File::Spec->catfile( $dir, 'package.js' );
+		if( -r $jspck ){
+			$res = {};
+			$res->{jspck} = $jspck;
+			my $content = path( $jspck )->slurp_utf8;
+			my @lines = split( /[\r\n]/, $content );
+			my $c;
+			# expect at least Package.describe() and maybe Package.onUse()
+			my $describe = $content;
+			$describe =~ s/^.*Package\.describe\s*\(\s*{([^\)}]+).*$/$1/s;
+			if( $describe ){
+				$res->{describe} = $describe;
+			} else {
+				msgWarn( "$jspck: 'Package.describe() call not found, so ignore the package" );
+				return undef;
+			}
+			# get the name (and the owner), making sure there is one and only one
+			my @names = grep( /\sname\s*:/, @lines );
+			$c = scalar( @names );
+			if( $c != 1 ){
+				msgWarn( "$jspck: found unexpected $c 'name' tag(s), so ignore the package" );
+				return undef;
+			} else {
+				my @parts = split( ':', $names[0], 2 );
+				my $name = $parts[1];
+				$name =~ s/^\s*//;
+				$name =~ s/,\s*$//;
+				$name =~ s/['"]//g;
+				my @words = split( ':', $name );
+				$res->{owner} = $words[0];
+				$res->{shortName} = $words[1];
+				$res->{name} = $name;
+			}
+			# get the version
+			my @versions = grep( /\sversion\s*:/, @lines );
+			$c = scalar( @versions );
+			if( $c != 1 ){
+				msgWarn( "$jspck: found unexpected $c 'version' tag(s), so ignore the package" );
+				return undef;
+			} else {
+				my @parts = split( ':', $versions[0], 2 );
+				my $version = $parts[1];
+				$version =~ s/^\s*//;
+				$version =~ s/,\s*$//;
+				$version =~ s/['"]//g;
+				$res->{version} = $version;
+			}
+			# get the summary
+			my @summaries = grep( /\ssummary\s*:/, @lines );
+			$c = scalar( @summaries );
+			if( $c != 1 ){
+				msgWarn( "$jspck: found unexpected $c 'summary' tag(s), so ignore the package" );
+				return undef;
+			} else {
+				my @parts = split( ':', $summaries[0], 2 );
+				my $summary = $parts[1];
+				$summary =~ s/^\s*//;
+				$summary =~ s/,\s*$//;
+				$summary =~ s/['"]//g;
+				$res->{summary} = $summary;
+			}
+			# get all api.use() dependents
+			# which can be specified as something.use( package ) or .use([ pack1, pck2 ]) with one package per line or all on the same line
+			# plus a small hack to consider the _use() cases
+			my @uses = ( $content =~ /[\._]use\s*\(([^\);]+)\)\s*;/g );
+			my $useres = {};
+			my $debug = false;
+			#$debug = true if $res->{name} eq 'accounts-iziam';
+			foreach my $it ( @uses ){
+				my $args = $it;
+				print "\$args ".Dumper( $args ) if $debug;
+				if( $args =~ m/^\s*\[/ ){
+					print "have [".EOL if $debug;
+					$args =~ s/^\s*\[([^\]]+)\].*$/$1/s;
+					print "\$args='$args'".EOL if $debug;
+					my @args = split( /[\r\n,]/, $args );
+					print "\@args ".Dumper( @args ) if $debug;
+					foreach my $arg ( @args ){
+						_getPackageDependency( $res, $useres, $arg );
+					}
+				} else {
+					print "do not have [".EOL if $debug;
+					my @args = split( /\s*,\s*/, $args, 2 );
+					$args = $args[0];
+					$args =~ s/^\s*([^,]+).*$/$1/;
+					_getPackageDependency( $res, $useres, $args );
+				}
+			}
+			my @useres = sort keys %{$useres};
+			$res->{uses} = \@useres;
+			# test other files
+			my $ok = true;
+			foreach my $it ( @{$Const->{package}{md}} ){
+				my $file = File::Spec->catfile( $dir, $it );
+				if( -r $file ){
+					$res->{$it} = $file;
+				} else {
+					msgWarn( "$file: not found or not readable" );
+					$ok = false;
+				}
+			}
+			return undef if !$ok;
+		} else {
+			msgVerbose( "$jspck: file not found or not readable" );
+			return undef;
+		}
+	} else {
+		msgVerbose( "$dir: not a directory" );
+		return undef;
+	}
+
+	return $res;
+}
+
+sub _getPackageDependency {
+	my ( $self, $deps, $arg ) = @_;
+
+	$arg =~ s/^\s*['",]\s*//g;
+	$arg =~ s/\s*['",]\s*$//g;
+	$arg =~ s/^\s*$//;
+	$arg =~ s/\@.*$//;
+
+	# do not consider as a dependency this same package
+	$deps->{$arg} = true if $arg && $arg ne "$self->{name}" && $arg !~ m/\.\.\.arguments/;
+}
+
+# ------------------------------------------------------------------------------------------------
+# get the list of .md files of a package
+# (I):
+# - none
+# (O):
+# - the list of MD files a package is expected to manage as an array ref
+
+sub getPackageMD {
+	return $Const->{package}{md};
 }
 
 1;